removed duplicated code, now one button class with enum types
