+++ app/src/main/java/com/example/di_1_hexentanz/GameBoard/CustomButtons/IButton.java
@@ -1,6 +1,7 @@
-package com.example.di_1_hexentanz.GameBoard.Button;
+package com.example.di_1_hexentanz.GameBoard.CustomButtons;
 
 import android.graphics.Bitmap;
+
 
 public interface IButton {
 
@@ -9,7 +10,10 @@ public interface IButton {
     int getLeftPosition();
     int getTopPosition();
 
-    Bitmap getBitmap();
+    Bitmap getBitmap(BtnType typeBtn);
 
+    enum BtnType{
+        YourTurnButton, YesButton, NoButton
+    }
 
 }
