+++ app/src/main/java/com/example/di_1_hexentanz/GameBoard/Gamescreen.java
@@ -20,9 +20,8 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.example.di_1_hexentanz.Dice.DiceUI;
-import com.example.di_1_hexentanz.GameBoard.Button.NoButton;
-import com.example.di_1_hexentanz.GameBoard.Button.YesIButton;
-import com.example.di_1_hexentanz.GameBoard.Button.YourTurnIButton;
+import com.example.di_1_hexentanz.GameBoard.CustomButtons.IButton;
+import com.example.di_1_hexentanz.GameBoard.CustomButtons.CustomButton;
 import com.example.di_1_hexentanz.Player;
 import com.example.di_1_hexentanz.PlayerColor;
 import com.example.di_1_hexentanz.R;
@@ -115,14 +114,14 @@ public class Gamescreen extends AppCompatActivity implements SensorEventListener
         drawBoardGame();
 
 
-        YourTurnIButton yourTurnButton = new YourTurnIButton(getApplicationContext(), displayMetrics);
+        CustomButton yourTurnButton = new CustomButton(getApplicationContext(), displayMetrics, IButton.BtnType.YourTurnButton);
         addContentView(yourTurnButton, findViewById(R.id.contraintLayout).getLayoutParams());
 
-        YesIButton yb = new YesIButton(getApplicationContext(), displayMetrics);
+        CustomButton yb = new CustomButton(getApplicationContext(), displayMetrics, IButton.BtnType.YesButton);
         addContentView(yb, findViewById(R.id.contraintLayout).getLayoutParams());
         yb.setVisibility(View.INVISIBLE);
 
-        NoButton nb = new NoButton(getApplicationContext(), displayMetrics);
+        CustomButton nb = new CustomButton(getApplicationContext(), displayMetrics, IButton.BtnType.NoButton);
         addContentView(nb, findViewById(R.id.contraintLayout).getLayoutParams());
         nb.setVisibility(View.INVISIBLE);
 
@@ -301,7 +300,7 @@ public class Gamescreen extends AppCompatActivity implements SensorEventListener
 
 
 
-    public void witchSelected(final Witch witch, YesIButton yb, NoButton nb) {
+    public void witchSelected(final Witch witch, CustomButton yb, CustomButton nb) {
         setState(GameState.ConfirmSelection);
         witch.getCurrentField().highlight();
         TextView outputtext = findViewById(R.id.TestDisplay);
@@ -323,7 +322,7 @@ public class Gamescreen extends AppCompatActivity implements SensorEventListener
         this.txtHome.setText("At home: " + n);
     }
 
-    public void putWitchOnGameboard(Witch witch, YesIButton yb, NoButton nb) {
+    public void putWitchOnGameboard(Witch witch, CustomButton yb, CustomButton nb) {
         Feld destination = felder[(witch.getPlayer().getStartFeld().getNumber() + lastDiceResult-1) % 36];
         witch.putWitchOnGameboard(this, destination);
         yb.setVisibility(View.INVISIBLE);
